requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confluent-kafka==2.3.0
cron-descriptor==1.4.3
Django==3.2.25
django-celery-beat==2.6.0
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
greenlet==3.0.3
gunicorn==22.0.0
inflection==0.5.1
kombu==5.3.7
loguru==0.7.2
packaging==24.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
psycopg2-binary==2.9.9
